+++ src/translations/en.php
@@ -1,0 +1,10 @@
+<?php
+
+return [
+    'An error occurred when trying to load this image' => 'An error occurred when trying to load this image',
+    'Click image to set focal point' => 'Click image to set focal point',
+    'Default Focal Point' => 'Default Focal Point',
+    'Max Thumb Height' => 'Max Thumb Height',
+    'Max Thumb Width' => 'Max Thumb Width',
+    'This field type can only be used on images' => 'This field type can only be used on images',
+];