+++ CHANGELOG.md
@@ -1,8 +1,25 @@
-# Focal Point Changelog
+# Focal Point Field Changelog
 
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## 2.0.0 - 2022-05-21
+
+### Added
+- Craft 4.0 compatibility
+- Added "Maximum thumb width" and "Maximum thumb height" field settings  
+
+### Changed
+- Changed name from "Focal Point" to "Focal Point Field"  
+- Changed package name from `vaersaagod/focalpoint` to `vaersaagod/focalpointfield`  
+- Changed plugin handle from `focal-point` to `focal-point-field`
+
+### Improved  
+- Focal point fields now work for any (image) extension
+
+### Fixed
+- Fixes an issue where it wasn't possible to add focal point fields to assets directly
 
 ## 1.0.3 - 2021-08-31
 ### Fixed